backend/requirements.txt
fastapi[all]==0.111.0
sqlalchemy[asyncio]==2.0.30
alembic==1.13.1
asyncpg==0.29.0
pydantic-settings==2.2.1
python-dotenv==1.0.1
uvicorn==0.30.1
pytest==8.2.1
pytest-asyncio==0.23.7
pytest-cov
httpx==0.27.0
python-multipart==0.0.9
faker
